+++ majorLink/src/main/java/com/example/majorLink/dto/response/NotificationResponse.java
@@ -1,10 +1,14 @@
 package com.example.majorLink.dto.response;
 
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @Getter
 @Builder
+@NoArgsConstructor
+@AllArgsConstructor
 public class NotificationResponse {
     private Long id;
     private String receiver;
